run experiments with building dataset

+++ src/trainer2.py
@@ -130,7 +130,7 @@ class Trainer:
             preds = torch.tensor(outputs, device='cuda')
 
         metric_fn.to('cuda')
-        metric_fn.update(preds, target)
+        metric_fn.update(preds, target.long())
 
         return loss
